% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.danger.R
\name{ds.danger}
\alias{ds.danger}
\title{Computes the statistical mean of a given vector}
\usage{
ds.danger(x = NULL, type = "combine", datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a numerical vector}

\item{type}{a character which represents the type of analysis to carry out.
If \code{type} is set to 'combine', a global mean is calculated
if \code{type} is set to 'split', the mean is calculated separately for each study.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
a numeric
}
\description{
This function is similar to the R function \code{mean}.
}
\details{


It is a wrapper for the server side function.
}
\author{
Gaye A., Isaeva I.
}
\seealso{
\code{ds.quantileMean} to compute quantiles.

\code{ds.summary} to generate the summary of a variable.
}

