% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DANGERdfEXTRACT.R
\name{ds.DANGERdfEXTRACT}
\alias{ds.DANGERdfEXTRACT}
\title{DANGER function returning data.frame (df) from processing servers to
clientside - *** DEPRECATED BY ds.DANGERdmtEXTRACT ***}
\usage{
ds.DANGERdfEXTRACT(
  dataframeName = NULL,
  extract.study.specific = TRUE,
  extract.all.studies.combined = TRUE,
  datasources = NULL
)
}
\arguments{
\item{dataframeName}{character string that specifies the name of the df 
to be copied from the serverside to the clientside}

\item{extract.study.specific}{logical defaulted to TRUE which means that the
df from each processing source will be combined and returned as a single object.
If FALSE this will not be returned. If the <extract.study.specific> and the
<extract.all.studies.combined> arguments are both TRUE separate and
combined dmts will be returned.}

\item{extract.all.studies.combined}{logical defaulted to TRUE which means that the
dfs from each processing source will be returned as a separate object. If FALSE
these will not be returned}

\item{datasources}{specifies the particular 'connection object(s)' to use.}
}
\value{
the serverside df specified by name dfName as a
data.frame/matrix/tibble on the clientside.
}
\description{
Copies a df from data servers to clientside.
}
\details{
*** DEPRECATED BY ds.DANGERdmtEXTRACT ***
}
\author{
DataSHIELD Development Team
}
