% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DANGERplot.R
\name{ds.DANGERplot}
\alias{ds.DANGERplot}
\title{DANGER function}
\usage{
ds.DANGERplot(x.name = NULL, y.name = NULL, z.name = NULL, datasources = NULL)
}
\arguments{
\item{x.name}{character string naming first object to be
plotted on clientside
from serverside to clientside}

\item{y.name}{character string naming second object to be
plotted on clientside}

\item{z.name}{character string naming third object to be 
plotted on clientside}

\item{datasources}{specifies the particular 'connection object(s)' to use.}
}
\value{
DANGER function
}
\description{
DANGER function
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
