% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DANGERseedEXTRACT.R
\name{ds.DANGERseedEXTRACT}
\alias{ds.DANGERseedEXTRACT}
\title{ds.DANGERseedEXTRACT}
\usage{
ds.DANGERseedEXTRACT(datasources = NULL)
}
\arguments{
\item{datasources}{specifies the particular 'connection object(s)' to use.}
}
\value{
DANGER function
}
\description{
DANGER function
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
