% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DANGERvarsEXTRACT.R
\name{ds.DANGERvarsEXTRACT}
\alias{ds.DANGERvarsEXTRACT}
\title{ds.DANGERvarsEXTRACT}
\usage{
ds.DANGERvarsEXTRACT(
  x.name = NULL,
  y.name = NULL,
  z.name = NULL,
  u.name = NULL,
  v.name = NULL,
  w.name = NULL,
  datasources = NULL
)
}
\arguments{
\item{x.name}{character string naming first variable to be copied
from serverside to clientside}

\item{y.name}{character string naming second variable to be copied
from serverside to clientside}

\item{z.name}{character string naming third variable to be copied
from serverside to clientside}

\item{u.name}{character string naming fourth variable to be copied
from serverside to clientside}

\item{v.name}{character string naming fifth variable to be copied
from serverside to clientside}

\item{w.name}{character string naming sixth variable to be copied
from serverside to clientside}

\item{datasources}{specifies the particular 'connection object(s)' to use.}
}
\value{
DANGER function
}
\description{
DANGER function
}
\details{
DANGER function
}
\author{
DataSHIELD Development Team
}
