% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.danger.R
\name{ds.danger}
\alias{ds.danger}
\title{Returns any object to the client side}
\usage{
ds.danger(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a server side object}
}
\value{
the server side object requested
}
\description{
DANGER - special function to allow analysts to look at server side objects
on the client side
}
\details{
this function is not meant to be used anywhere except in a development environment
It is designed to give the analyst access to the data (i.e. a route through the DataSHIELD
defences)
}
\author{
Bishop T.
}
